requirements.txt
pytest==9.0.2
allure-pytest==2.15.3
selenium==4.27.0
selene==2.0.0rc6
